move is_scalar up
